test: split file-ops into unit and integration tests
